runner.rsを6つのサブモジュールに分割 #75
runner.rs (1893行) の実行ロジックを機能単位のサブモジュールに分割する。

- runner/note.rs      音符と休符の実行
- runner/tie.rs       タイ(&)の各モードの処理
- runner/function.rs  ユーザー定義関数・システム関数・変数展開
- runner/control.rs   条件分岐とループの実行
- runner/cc.rs        コントロールチェンジ・テンポ・音色
- runner/structure.rs サブルーチン・連符・和音・タイムポインタ

lexer の分割と同じ方針で、src/runner.rs を親モジュールとして残し
src/runner/*.rs を追加した (mod.rs へのリネームはしない)。
移動した関数の可視性は pub(super) とした。

trk! マクロは macro_rules! のテキストスコープに従うため、
子モジュールから利用できるようマクロ定義の直後に mod 宣言を置いている。

exec() 本体、LoopItem、および pub 関数 (exec_args / exec_value /
exec_value_int / exec_value_int_by_token / value_range / exec_easy) は
runner.rs に残したため、公開APIは変更していない。

このコミットは移動のみで振る舞いは変更していない。
cargo test 102件通過、ゴールデンMIDIのハッシュも一致。
分割前後の全行を正規化して照合し完全一致を確認済み。
