+++ src/runner.rs
@@ -33,6 +33,20 @@ macro_rules! trk {
     };
 }
 
+mod cc;
+mod control;
+mod function;
+mod note;
+mod structure;
+mod tie;
+
+use cc::*;
+use control::*;
+use function::*;
+use note::*;
+use structure::*;
+use tie::*;
+
 /// run tokens and get arguments(=`Vec<Token>`)
 pub fn exec_args(song: &mut Song, tokens: &Vec<Token>) -> Vec<SValue> {
     let mut args: Vec<SValue> = vec![];
@@ -68,7 +82,6 @@ pub fn exec_value_int_by_token(song: &mut Song, tok: &Token) -> isize {
     let tokens = tok.children.as_ref().unwrap_or(&empty_tokens);
     exec_value_int(song, tokens)
 }
-
 
 /// MetaTextに書き込める文字列は127バイトまでなので、文字境界を保ったまま切り詰める
 fn trim_meta_text(txt_raw: &str) -> String {
@@ -939,7 +952,6 @@ pub fn exec(song: &mut Song, tokens: &Vec<Token>) -> bool {
     true
 }
 
-
 fn runtime_error(song: &mut Song, msg: &str) {
     song.add_log(format!(
         "[ERROR]({}) {}: {}",
@@ -947,931 +959,6 @@ fn runtime_error(song: &mut Song, msg: &str) {
         song.get_message(MessageKind::RuntimeError),
         msg
     ));
-}
-
-fn exec_play(song: &mut Song, t: &Token) -> bool {
-    let tmp_cur_track = song.cur_track;
-    let lineno = t.lineno;
-    let start_pos = trk!(song).timepos;
-    let mut time_ptr_last = start_pos;
-    // play
-    let arg_tokens = t.children.clone().unwrap_or(vec![]);
-    for (index, arg) in arg_tokens.iter().enumerate() {
-        song.change_cur_track(index + 1);
-        trk!(song).timepos = start_pos;
-        // eval calc
-        let src = exec_value(song, &vec![arg.clone()]).to_s();
-        // println!("play(TR={})({}):{}", index+1, lineno, src);
-        // eval tokens
-        let tokens = lex(song, &src, lineno);
-        exec(song, &tokens);
-        // check lastpos
-        if trk!(song).timepos > time_ptr_last { time_ptr_last = trk!(song).timepos; }
-    }
-    song.track_sync();
-    song.cur_track = tmp_cur_track;
-    true
-}
-
-fn exec_cc_rpn_nrpn(song: &mut Song, t: &Token, cc1: isize, cc2: isize, cc3: isize) {
-    let val = exec_value_int_by_token(song, t);
-    let msb = t.data[0].to_i();
-    let lsb = t.data[1].to_i();
-    song.add_event(Event::cc(trk!(song).timepos, trk!(song).channel, cc1, msb));
-    song.add_event(Event::cc(trk!(song).timepos, trk!(song).channel, cc2, lsb));
-    song.add_event(Event::cc(trk!(song).timepos, trk!(song).channel, cc3, val)); 
-}
-
-fn exec_cc_rpn_nrpn_direct(song: &mut Song, t: &Token, cc1: isize, cc2: isize, cc3: isize) {
-    let args = exec_args(song, t.children.as_ref().unwrap_or(&vec![]));
-    if args.len() != 3 {
-        runtime_error(song, "RPN/NRPN needs 3 arguments");
-        return;
-    }
-    let msb = args[0].to_i();
-    let lsb = args[1].to_i();
-    let val = args[2].to_i();
-    song.add_event(Event::cc(trk!(song).timepos, trk!(song).channel, cc1, msb));
-    song.add_event(Event::cc(trk!(song).timepos, trk!(song).channel, cc2, lsb));
-    song.add_event(Event::cc(trk!(song).timepos, trk!(song).channel, cc3, val)); 
-}
-
-fn exec_userfunc_or_array_or_macro(song: &mut Song, t: &Token) -> bool {
-    // check value is array?
-    if t.data.len() > 0 {
-        let name = t.data[0].to_s();
-        let var = song.variables_get(&name).unwrap_or(&SValue::None).clone();
-        match var {
-            // is Array
-            SValue::Array(a) => {
-                // get arg
-                let args_tokens = t.children.clone().unwrap();
-                let args: Vec<SValue> = exec_args(song, &args_tokens);
-                if args.len() == 0 {
-                    runtime_error(song, &format!("get Array({}) element needs arguments", name));
-                    return false;
-                }
-                let index = args[0].to_i() as usize;
-                if a.len() <= index {
-                    runtime_error(song, &format!("Array({}) index out of range", name));
-                    return false;
-                }
-                let v = a[index].clone();
-                song.stack.push(v);
-                return true;
-            },
-            // is String macro >>> exec string
-            SValue::Str(src, _) => {
-                // get arg
-                let args_tokens = t.children.clone().unwrap();
-                let args: Vec<SValue> = exec_args(song, &args_tokens);
-                if args.len() == 0 {
-                    runtime_error(song, &format!("get String({}) element needs arguments", name));
-                    return false;
-                }
-                // replace string
-                let mut s = src.clone();
-                for (i, v) in args.iter().enumerate() {
-                    let varname = format!("#?{}", i + 1);
-                    s = s.replace(&varname, &v.to_s());
-                }
-                let tokens = lex(song, &s, t.lineno);
-                return exec(song, &tokens);
-            },
-            _ => {}
-        }
-    }
-    // check func_id
-    let func_id = t.tag as usize;
-    if song.functions.len() <= func_id {
-        runtime_error(song, &format!("broken func_id={} in exec_call_user_function", func_id));
-        return false;
-    }
-    // println!("call_user_function[{}]={}::{:?}", func_id, song.functions[func_id].name, song.functions[func_id].arg_def_values);
-    
-    song.variables_stack_push();
-    // eval args
-    let args_tokens = t.children.clone().unwrap();
-    let args: Vec<SValue> = exec_args(song, &args_tokens);
-    // set local variables
-    for i in 0..song.functions[func_id].arg_names.len() {
-        let varname = &song.functions[func_id].arg_names[i].clone();
-        let mut v: SValue = if i < args.len() { args[i].clone() } else { SValue::None };
-        v = match v {
-            SValue::None => song.functions[func_id].arg_def_values[i].clone(),
-            _ => v,
-        };
-        song.variables_insert(varname, v);
-    }
-    // eval function
-    let tokens = song.functions[func_id].tokens.clone();
-    let tmp_break_flag = song.flags.break_flag;
-    // println!("func_body={:?}", tokens);
-    let eval_result = exec(song, &tokens);
-    song.flags.break_flag = tmp_break_flag;
-    let vars = song.variables_stack_pop();
-    if song.flags.function_needs_return_value {
-        let return_val = vars.get("Result");
-        song.stack.push(return_val.unwrap_or(&SValue::None).clone());
-    }
-    eval_result
-}
-
-fn exec_sys_function(song: &mut Song, t: &Token) -> bool {
-    let args_tokens = t.children.clone().unwrap_or(vec![]);
-    let args:Vec<SValue> = exec_args(song, &args_tokens);
-    let func_name = if t.data.len() > 0 { t.data[0].to_s() } else { "".to_string() };
-    // is user function?
-    let func_val = song.variables_get(&func_name).unwrap_or(&SValue::new()).clone();
-    match func_val {
-        SValue::UserFunc(_func_id) => {
-            if exec_userfunc_or_array_or_macro(song, t) { return true; }
-        },
-        _ => {}, // maybe system function
-    }
-    //
-    // todo: https://sakuramml.com/wiki/index.php?%E7%B5%84%E3%81%BF%E8%BE%BC%E3%81%BF%E9%96%A2%E6%95%B0
-    //
-    // 参照できるシステム関数
-    if let Some(cb) = song.calc_functions.get(&func_name) {
-        let result = cb(song, args);
-        song.stack.push(result);
-    } else {
-        // macro ("=var_name")
-        let func_name2 = if func_name.len() >= 2 { func_name[1..].to_string() } else { func_name };
-        let args = t.children.clone().unwrap_or(vec![]);
-        let args = exec_args(song, &args);
-        let val = song.variables_get(&func_name2).unwrap_or(&SValue::new()).clone();
-        let mut val_s = val.to_s();
-        for (index, arg) in args.iter().enumerate() {
-            let macro_n = format!("#?{}", index+1);
-            let target = arg.clone().to_s();
-            val_s = val_s.replace(&macro_n, &target);
-        }
-        // println!("macro={}//{:?}", val_s, t);
-        if song.flags.function_needs_return_value {
-            song.stack.push(SValue::from_s(val_s));
-        } else {
-            // exec macro
-            let tokens = lex(song, &val_s, t.lineno);
-            exec(song, &tokens);
-        }
-    }
-    true
-}
-
-fn exec_if(song: &mut Song, t: &Token) -> bool {
-    let children = match &t.children {
-        Some(tokens) => tokens,
-        None => return false,
-    };
-    if children.len() < 3 {
-        return false;
-    }
-    let cond_token = &children[0];
-    let true_token = &children[1];
-    let false_token = &children[2];
-    // eval cond
-    let cond = cond_token.children.clone().unwrap();
-    let cond_val = exec_value(song, &cond);
-    // exec true or false
-    if cond_val.to_i() != 0 {
-        let tokens = true_token.children.clone().unwrap();
-        exec(song, &tokens);
-    } else {
-        let tokens = false_token.children.clone().unwrap();
-        exec(song, &tokens);
-    }
-    true
-}
-
-fn exec_while(song: &mut Song, t: &Token) -> bool {
-    let children = match &t.children {
-        Some(tokens) => tokens,
-        None => return false,
-    };
-    if children.len() < 2 {
-        return false;
-    }
-    let cond_token = &children[0];
-    let body_token = &children[1];
-    let mut counter = 0;
-    // loop
-    loop {
-        // eval cond
-        let cond = cond_token.children.clone().unwrap();
-        let cond_val = exec_value(song, &cond);
-        if cond_val.to_b() == false {
-            break;
-        }
-        // exec body
-        let body = body_token.children.clone().unwrap();
-        exec(song, &body);
-        // check counter
-        counter += 1;
-        if counter > song.flags.max_loop {
-            song.add_log(format!(
-                "[ERROR]({}) {} WHILE(>{})", t.lineno, 
-                song.get_message(MessageKind::LoopTooManyTimes),
-                song.flags.max_loop
-            ));
-            break;
-        }
-        // check break flag
-        match song.flags.break_flag {
-            1 => {
-                song.flags.break_flag = 0;
-                break;
-            },
-            2 => {
-                song.flags.break_flag = 0;
-                continue;
-            },
-            3 => {
-                break;
-            }
-            _ => {},
-        }
-    }
-    true
-}
-
-fn exec_for(song: &mut Song, t: &Token) -> bool {
-    let children = match &t.children {
-        Some(tokens) => tokens,
-        None => return false,
-    };
-    if children.len() < 4 {
-        return false;
-    }
-    let init_token = &children[0];
-    let cond_token = &children[1];
-    let inc_token = &children[2];
-    let body_token = &children[3];
-    // eval init
-    let init = init_token.children.clone().unwrap();
-    exec(song, &init);
-    let mut counter = 0;
-    // loop
-    loop {
-        // eval cond
-        let cond = cond_token.children.clone().unwrap();
-        let cond_val = exec_value(song, &cond);
-        if cond_val.to_b() == false {
-            break;
-        }
-        // exec body
-        let body = body_token.children.clone().unwrap();
-        exec(song, &body);
-        // check loop counter
-        counter += 1;
-        if counter > song.flags.max_loop {
-            song.add_log(format!(
-                "[ERROR]({}) {} FOR(>{})", t.lineno, 
-                song.get_message(MessageKind::LoopTooManyTimes),
-                song.flags.max_loop
-            ));
-            break;
-        }
-        // inc
-        let inc_tokens = inc_token.children.clone().unwrap();
-        // check break or continue
-        if song.flags.break_flag == 1 { // break
-            song.flags.break_flag = 0;
-            break;
-        }
-        if song.flags.break_flag == 2 { // continue
-            song.flags.break_flag = 0;
-            exec(song, &inc_tokens); // eval inc
-            continue;
-        }
-        // eval inc
-        exec(song, &inc_tokens); // eval inc
-    }
-    true
-}
-
-fn get_system_value(cmd: &str, song: &Song) -> Option<SValue> {
-    // <SYSTEM_REF>
-    if cmd == "TR" || cmd == "TRACK" || cmd == "Track" { // @ get current track no - 現在のトラック番号を得る
-        let tr = song.cur_track as isize;
-        return Some(SValue::from_i(tr));
-    }
-    if cmd == "CH" || cmd == "CHANNEL" { // @ get current channel no - 現在のチャンネル番号を得る
-        let ch = trk!(song).channel + 1; // range: 1-16
-        return Some(SValue::from_i(ch));
-    }
-    if cmd == "TIME" || cmd == "Time" || cmd == "TIMEPOS" || cmd == "TIMEPTR" { // @ get time posision - 現在のタイムポインタ値を得る
-        let v = trk!(song).timepos;
-        return Some(SValue::from_i(v));
-    }
-    if cmd == "TEMPO" || cmd == "Tempo" || cmd == "BPM" { // @ get tempo - 現在のテンポ値を得る
-        let v = song.tempo;
-        return Some(SValue::from_i(v));
-    }
-    if cmd == "KEY" || cmd == "KEY_SHIFT" { // @ get key shift - 現在のキーシフト値を得る
-        let v = song.key_shift;
-        return Some(SValue::from_i(v));
-    }
-    if cmd == "TR_KEY" || cmd == "TrackKey" { // @ get track key shift - 現在のトラックごとのキーシフト値を得る
-        let v = trk!(song).track_key;
-        return Some(SValue::from_i(v));
-    }
-    if cmd == "TIMEBASE" || cmd == "Timebase" { // @ get timebase - 現在のタイムベース値を得る
-        let v = song.timebase;
-        return Some(SValue::from_i(v));
-    }
-    if cmd == "l" { // @ get length - 現在のlの値を得る
-        let v = trk!(song).length;
-        return Some(SValue::from_i(v));
-    }
-    if cmd == "v" { // @ get velocity - 現在のvの値を得る
-        let v = trk!(song).velocity;
-        return Some(SValue::from_i(v));
-    }
-    if cmd == "q" { // @ get gate rate - 現在のqの値を得る
-        let v = trk!(song).qlen;
-        return Some(SValue::from_i(v));
-    }
-    if cmd == "o" { // @ get octave rate - 現在のoの値を得る
-        let v = trk!(song).octave;
-        return Some(SValue::from_i(v));
-    }
-    // </SYSTEM_REF>
-    None
-}
-
-fn var_extract(val: &SValue, song: &mut Song) -> SValue {
-    match val {
-        // String
-        SValue::Str(s, _) => {
-            if s.starts_with('=') && s.len() >= 2 {
-                let key = &s[1..];
-                match song.variables_get(key) {
-                    Some(v) => v.clone(),
-                    None => {
-                        match get_system_value(key, song) {
-                            Some(v) => return v,
-                            None => {
-                                let err_msg = format!("[WARN]({}) Undefined: {}", song.lineno, key);
-                                song.add_log(err_msg);
-                                SValue::None
-                            },
-                        }
-                    }
-                }
-            } else {
-                SValue::from_str(&s)
-            }
-        },
-        // Other value
-        _ => val.clone(),
-    }
-}
-
-fn tempo_change_a_to_b(song: &mut Song, a: isize, b: isize, len: isize) {
-    let step = (song.timebase * 4) / 16;
-    let step_cnt = len / step;
-    let width = b - a;
-    let timepos = trk!(song).timepos;
-    for i in 0..step_cnt {
-        let v = (a as f32) + (width as f32) * (i as f32 / step_cnt as f32);
-        tempo_change(song, v as isize);
-        trk!(song).timepos += step;
-    }
-    trk!(song).timepos = timepos + len;
-    tempo_change(song, b);
-    trk!(song).timepos = timepos;
-}
-
-fn tempo_change(song: &mut Song, tempo: isize) {
-    song.tempo = tempo;
-    let mpq = if tempo > 0 { 60000000 / tempo } else { 120 };
-    let e = Event::meta(
-        trk!(song).timepos,
-        0xFF,
-        0x51,
-        0x03,
-        vec![
-            (mpq >> 16 & 0xFF) as u8,
-            (mpq >> 8 & 0xFF) as u8,
-            (mpq >> 0 & 0xFF) as u8,
-        ],
-    );
-    song.add_event(e);
-}
-
-fn exec_sub(song: &mut Song, t: &Token) {
-    let timepos_tmp: isize;
-    {
-        let trk = &song.tracks[song.cur_track];
-        timepos_tmp = trk.timepos;
-    }
-    {
-        let _ = match &t.children {
-            Some(tokens) => exec(song, tokens),
-            None => false,
-        };
-    }
-    {
-        let trk = &mut song.tracks[song.cur_track];
-        trk.timepos = timepos_tmp;
-    }
-}
-
-fn exec_div(song: &mut Song, t: &Token) {
-    let len_s = &t.data[0].to_s();
-    let cnt = t.value_i;
-    let length_org: isize;
-    let timepos_end: isize;
-    {
-        let trk = &mut song.tracks[song.cur_track];
-        let div_len = calc_length(len_s, song.timebase, trk.length);
-        let note_len = if cnt > 0 { div_len / cnt } else { 0 };
-        timepos_end = trk.timepos + div_len;
-        length_org = trk.length;
-        trk.length = note_len;
-    }
-    let _ = match &t.children {
-        None => false,
-        Some(tokens) => exec(song, tokens),
-    };
-    // clean
-    {
-        let trk = &mut song.tracks[song.cur_track];
-        trk.timepos = timepos_end;
-        trk.length = length_org;
-    }
-}
-
-fn exec_harmony(song: &mut Song, t: &Token, flag_begin: bool) {
-    // begin
-    if flag_begin {
-        song.flags.harmony_flag = true;
-        song.flags.harmony_time = song.tracks[song.cur_track].timepos;
-        return;
-    }
-    // end
-    if song.flags.harmony_flag {
-        song.flags.harmony_flag = false;
-        // get harmony length
-        let note_len_s = t.data[0].to_s();
-        let mut note_qlen = t.data[1].to_i();
-        let note_vel = t.data[2].clone();
-        // parameters
-        if note_qlen < 0 {
-            note_qlen = trk!(song).qlen;
-        }
-        let note_len = calc_length(&note_len_s, song.timebase, trk!(song).length);
-        // change event length
-        while song.flags.harmony_events.len() > 0 {
-            let mut e = song.flags.harmony_events.pop().unwrap();
-            e.time = song.flags.harmony_time;
-            if note_qlen != 0 {
-                e.v2 = note_len * note_qlen / 100;
-            }
-            if !note_vel.is_none() {
-                e.v3 = note_vel.to_i();
-            }
-            trk!(song).events.push(e);
-        }
-        trk!(song).timepos = song.flags.harmony_time + note_len;
-        return;
-    }
-}
-
-fn exec_get_time(song: &mut Song, t: &Token, cmd: &str) -> isize{
-    // Calc Time (SakuraObj_time2step)
-    // (ref) https://github.com/kujirahand/sakuramml-c/blob/68b62cbc101669211c511258ae1cf830616f238e/src/k_main.c#L473
-    let args = exec_args(song, t.children.as_ref().unwrap_or(&Vec::new()));
-    if args.len() == 0 {
-        runtime_error(song, &format!("[{}] no arguments", cmd));
-        return 0;
-    }
-    if args.len() == 1 {
-        return args[0].to_i();
-    }
-    if args.len() < 3 {
-        runtime_error(song, &format!("[{}] needs 1 or 3 arguments", cmd));
-        return 0;
-    }
-    let mes = args[0].to_i() + song.flags.measure_shift;
-    let beat = args[1].to_i();
-    let tick = args[2].to_i();
-
-    // calc
-    let base = song.timebase * 4 / song.timesig_deno;
-    let total = (mes - 1) * (base * song.timesig_frac) + (beat - 1) * base + tick;
-    total
-}
-
-fn get_note_info_from_token(t: &Token) -> NoteInfo {
-    let data = &t.data;
-    if data.len() < 8 { // broken note
-        return NoteInfo {
-            no: 0,
-            flag: 0,
-            natural: 0,
-            len_s: "".to_string(),
-            qlen: 0,
-            vel: 0,
-            t: 0,
-            o: 0,
-            slur: 0,
-        };
-    }
-    let note_no = (t.value_i % 12) as isize;
-    let data_note_flag = data[0].to_i();
-    let data_note_natural = data[1].to_i();
-    let data_note_len = data[2].to_s();
-    let data_note_qlen = data[3].to_i(); // 0
-    let data_note_vel = data[4].to_i(); // -1
-    let data_note_t = data[5].to_i(); // isize::MIN
-    let data_note_o = data[6].to_i(); // -1
-    let data_slur = data[7].to_i(); // 0 or 1 --- TODO: #7
-    NoteInfo {
-        no: note_no,
-        flag: data_note_flag,
-        natural: data_note_natural,
-        len_s: data_note_len,
-        qlen: data_note_qlen,
-        vel: data_note_vel,
-        t: data_note_t,
-        o: data_note_o,
-        slur: data_slur,
-    }
-}
-
-/// get note info, and shift key
-fn set_note_info_with_default_value(note: &mut NoteInfo, song: &mut Song) {
-    // set note with track's default value
-    if note.qlen == 0 {
-        note.qlen = trk!(song).qlen;
-    }
-    if note.vel < 0 {
-        note.vel = trk!(song).velocity;
-    }
-    if note.t == isize::MIN {
-        note.t = trk!(song).timing;
-    }
-    if note.o < 0 {
-        note.o = trk!(song).octave;
-    }
-    // calc note no
-    let mut noteno = note.o * 12 + note.no + note.flag;
-    // key_shift / key_flag / track_key
-    if song.use_key_shift {
-        noteno += if note.natural == 0 {
-            song.key_flag[(note.no as usize) % 12]
-        } else {
-            0
-        };
-        noteno += song.key_shift;
-        noteno += trk!(song).track_key;
-    }
-    note.no = noteno;
-}
-
-fn exec_note(song: &mut Song, t: &Token) {
-    // get note parameters
-    let mut note = get_note_info_from_token(t);
-    // get note info, and shift key
-    set_note_info_with_default_value(&mut note, song);
-    // timepos
-    let timepos = trk!(song).timepos;
-    let start_pos = timepos;
-    // onTime / onNote
-    let v = trk!(song).calc_v_on_time(note.vel);
-    let v = trk!(song).calc_v_on_note(v);
-    let t = trk!(song).calc_t_on_note(note.t);
-    let qlen = trk!(song).calc_qlen_on_note(note.qlen);
-    let o_abs = trk!(song).calc_o_on_note(-1);
-    if o_abs != -1 {// ノートはそのままでオクターブだけ変える
-        note.no = note.no % 12 + o_abs * 12; // set absolute octave
-    }
-    // Random
-    if trk!(song).o_rand > 0 { // octave randomize
-        let r = song.calc_rand_value(0, trk!(song).o_rand);
-        if r != 0 {
-            note.no += r * 12;
-        }
-    }
-    let v = if trk!(song).v_rand > 0 {
-        song.calc_rand_value(v, trk!(song).v_rand)
-    } else {
-        v
-    };
-    let t = if trk!(song).t_rand > 0 {
-        song.calc_rand_value(t, trk!(song).t_rand)
-    } else {
-        t
-    };
-    let qlen = if trk!(song).q_rand > 0 {
-        song.calc_rand_value(qlen, trk!(song).q_rand)
-    } else {
-        qlen
-    };
-    // note len
-    let mut notelen = calc_length(&note.len_s, song.timebase, trk!(song).length);
-    // note len onNote / onCycle
-    let notelen_on_note = trk!(song).calc_l_on_note(-1);
-    if notelen_on_note != -1 { // onNote / onCycle の値があれば強制的に上書き
-        notelen = notelen_on_note;
-    }
-    let notelen_real = (notelen as f32 * qlen as f32 / 100.0) as isize;
-    // check range
-    let v = value_range(0, v, 127);
-    // event
-    let event = Event::note(timepos + t, trk!(song).channel, note.no, notelen_real, v);
-    // println!("- {}: note(no={},len={},qlen={},v={},t={},o={})", trk.timepos, noteno, notelen_real, qlen, v, t, o);
-    trk!(song).timepos += notelen;
-
-    // octave_once?
-    if song.flags.octave_once != 0 {
-        trk!(song).octave = trk!(song).octave - song.flags.octave_once;
-        song.flags.octave_once = 0;
-    }
-
-    // harmony?
-    if song.flags.harmony_flag {
-        trk!(song).timepos = song.flags.harmony_time;
-        song.flags.harmony_events.push(event);
-        return;
-    }
-    // tie or slur?
-    if note.slur >= 1 {
-        trk!(song).tie_notes.push(event);
-        return;
-    }
-    if trk!(song).tie_notes.len() > 0 {
-        trk!(song).tie_notes.push(event);
-        check_tie_notes(song);
-        return;
-    }
-    // onNote event
-    trk!(song).write_cc_on_note(start_pos);
-    // onNoteWave event
-    trk!(song).write_cc_on_note_wave(start_pos);
-    // write note event
-    trk!(song).events.push(event);
-}
-
-/// TieMode::Port
-fn tie_mode_port(song: &mut Song) {
-    let mut last_note = trk!(song).tie_notes.remove(0);
-    let mut tie_value = trk!(song).tie_value;
-    loop {
-        if trk!(song).tie_notes.len() == 0 {
-            trk!(song).events.push(last_note);
-            break;
-        }
-        let next_event = trk!(song).tie_notes.remove(0);
-        // same note no
-        if last_note.v1 == next_event.v1 {
-            // add note length
-            let time_pos = next_event.time + next_event.v2;
-            last_note.v2 = time_pos - last_note.time;
-            continue;
-        }
-        // check bend range in track
-        let mut bend_range = trk!(song).bend_range;
-        if bend_range <= 0 {
-            // set bend range
-            trk!(song).bend_range = 12;
-            let timepos = if last_note.time <= 0 { 0 } else { last_note.time - 1 };
-            let bend_range_event = Event::pitch_bend_range(timepos, trk!(song).channel, 12);
-            trk!(song).events.push(bend_range_event);
-            bend_range = 12;
-        }
-        // calc pitch range
-        // bend value range: -8192 to 8191
-        let note_diff: isize = next_event.v1 - last_note.v1;
-        tie_value = if tie_value == 0 { (song.timebase * 4) / 8 } else { tie_value };
-        let bend_from = (note_diff as f32 * (8192f32 / bend_range as f32)) as isize;
-        let bend_to = 0;
-        let mut last_v = 0;
-        for i in 0..tie_value {
-            let timepos = next_event.time - tie_value + i;
-            let v = ((bend_from - bend_to) as f32 * (i as f32 / tie_value as f32)) as isize;
-            if last_v == v { continue; }
-            last_v = v;
-            let bend_event = Event::pitch_bend(timepos, trk!(song).channel, v + 8192);
-            trk!(song).events.push(bend_event);
-        }
-        last_note.v2 = next_event.time - last_note.time;
-        trk!(song).events.push(last_note);
-        let bend_event_end = Event::pitch_bend(next_event.time, trk!(song).channel, bend_to + 8192);
-        trk!(song).events.push(bend_event_end);
-        last_note = next_event;
-    }
-}
-
-fn tie_mode_bend(song: &mut Song) {
-    // first note
-    let mut last_note = trk!(song).tie_notes.remove(0);
-    let mut begin_note = last_note.clone();
-    // set bend range
-    let mut bend_range = trk!(song).bend_range;
-    if bend_range <= 0 {
-        trk!(song).bend_range = 12;
-        let timepos = if last_note.time <= 0 { 0 } else { last_note.time - 1 };
-        let bend_range_event = Event::pitch_bend_range(timepos, trk!(song).channel, 12);
-        trk!(song).events.push(bend_range_event);
-        bend_range = 12;
-    }
-    // set bend 0
-    let bend0 = Event::pitch_bend(last_note.time, trk!(song).channel, 8192);
-    trk!(song).events.push(bend0);
-    let mut lastpos = last_note.time + last_note.v2;
-    while trk!(song).tie_notes.len() > 0 {
-        let next_event = trk!(song).tie_notes.remove(0);
-        lastpos = next_event.time + next_event.v2;
-        // same note no
-        if last_note.v1 == next_event.v1 {
-            // add note length
-            let time_pos = next_event.time + next_event.v2;
-            last_note.v2 = time_pos - last_note.time;
-            continue;
-        }
-        // calc pitch range
-        // bend value range: -8192 to 8191
-        let note_diff: isize = next_event.v1 - last_note.v1;
-        let bend_event = Event::pitch_bend(
-            next_event.time,
-            trk!(song).channel,
-            (note_diff as f32 * 8192f32 / bend_range as f32) as isize + 8192,
-        );
-        trk!(song).events.push(bend_event);
-    }
-    // write begin note
-    begin_note.v2 = lastpos - begin_note.time;
-    trk!(song).events.push(begin_note);
-    // reset bend
-    let bend_end = Event::pitch_bend(lastpos, trk!(song).channel, 8192);
-    trk!(song).events.push(bend_end);
-}
-
-fn tie_mode_gate(song: &mut Song) {
-    let mut last_note = trk!(song).tie_notes.remove(0);
-    let tie_value = trk!(song).tie_value;
-    loop {
-        if trk!(song).tie_notes.len() == 0 {
-            trk!(song).events.push(last_note);
-            break;
-        }
-        let next_event = trk!(song).tie_notes.remove(0);
-        // same note no
-        if last_note.v1 == next_event.v1 {
-            // add note length
-            let time_pos = next_event.time + next_event.v2;
-            last_note.v2 = time_pos - last_note.time;
-            continue;
-        }
-        // different note no
-        if tie_value == 0 {
-            last_note.v2 = next_event.time - last_note.time;
-        } else {
-            last_note.v2 = tie_value;
-        }
-        trk!(song).events.push(last_note);
-        last_note = next_event;
-    }
-}
-
-/// alpeggio mode
-fn tie_mode_alpe(song: &mut Song) {
-    let last_note = &trk!(song).tie_notes[trk!(song).tie_notes.len() - 1];
-    let last_pos = last_note.time + last_note.v2;
-    let tie_notes = trk!(song).tie_notes.clone();
-    for mut event in tie_notes.into_iter() {
-        event.v2 = last_pos - event.time;
-        trk!(song).events.push(event);
-    }
-}
-
-fn check_tie_notes(song: &mut Song) {
-    // Tie/Slur mode (https://sakuramml.com/doc/command/11.htm)
-    //
-    // "Slur(type, value)"で、タイ記号"&(value)"の異音程(スラー)の動作を変更する。
-    // type	typeの概略	動作	valueの意味	rangeの意味	使い方例
-    // 0	グリッサンド	異音程をベンドで滑らかにつなぐ(※1)(※2)（※3）	グリッサンドの長さ	ベンドレンジを指定。省略可。	@81 l4 Slur(0,!8) c&e&g
-    // 1	ベンド	異音程をベンドで表現。ギターのハンマリングに近い。(※1)	無効	無効	@25 l8 Slur(1,0) cdc c&d&c g&f&e&d
-    // 2	ゲート	＆のついた音符のゲートを、valueにする	ゲートの長さ	無効	@81 l8 Slur(2,100) q50 c&d e&f g&f e&d
-    // 3	アルペジオ	＆でつないだ音符の終わりまでゲートを伸ばす。そのとき、value に、ノートの最大発音音数を指定できる。	最大発音音数	無効	l16 Slur(3,100) c&e&g d&f&a
-    if trk!(song).tie_notes.len() == 0 {
-        return;
-    }
-    match trk!(song).tie_mode {
-        TieMode::Port => tie_mode_port(song),
-        TieMode::Bend => tie_mode_bend(song),
-        TieMode::Gate => tie_mode_gate(song),
-        TieMode::Alpe => tie_mode_alpe(song),
-    };
-}
-
-fn exec_note_n(song: &mut Song, t: &Token) {
-    // parameters
-    let data_note_no = var_extract(&t.data[0], song).to_i();
-    let data_note_len = var_extract(&t.data[1], song).to_s();
-    let data_note_qlen = var_extract(&t.data[2], song).to_i(); // 0
-    let data_note_vel = var_extract(&t.data[3], song).to_i(); // -1
-    let data_note_t = var_extract(&t.data[4], song).to_i(); // isize::MIN
-    let start_pos = trk!(song).timepos;
-    let track_key = trk!(song).track_key;
-    let key_shift = song.key_shift;
-
-    // check parameters
-    let notelen = calc_length(&data_note_len, song.timebase, trk!(song).length);
-    let qlen = if data_note_qlen != 0 {
-        data_note_qlen
-    } else {
-        trk!(song).qlen
-    };
-    let v = if data_note_vel >= 0 {
-        data_note_vel
-    } else {
-        trk!(song).velocity
-    };
-    let t = if data_note_t != isize::MIN {
-        data_note_t
-    } else {
-        trk!(song).timing
-    };
-    // onTime / onNote
-    let v = trk!(song).calc_v_on_time(v);
-    let v = trk!(song).calc_v_on_note(v);
-    let t = trk!(song).calc_t_on_note(t);
-    let qlen = trk!(song).calc_qlen_on_note(qlen);
-    // Random
-    let v = if trk!(song).v_rand > 0 {
-        song.calc_rand_value(v, trk!(song).v_rand)
-    } else {
-        v
-    };
-    let t = if trk!(song).t_rand > 0 {
-        song.calc_rand_value(t, trk!(song).t_rand)
-    } else {
-        t
-    };
-    let qlen = if trk!(song).q_rand > 0 {
-        song.calc_rand_value(qlen, trk!(song).q_rand)
-    } else {
-        qlen
-    };
-    // calc
-    let notelen_real = (notelen as f32 * qlen as f32 / 100.0) as isize;
-    // range
-    let v = value_range(0, v, 127);
-    let event = Event::note(
-        trk!(song).timepos + t,
-        trk!(song).channel,
-        data_note_no + track_key + key_shift,
-        notelen_real,
-        v,
-    );
-    // println!("- {}: note(no={},len={},qlen={},v={},t={})", trk!(song).timepos, notelen_real, notelen, qlen, v, t);
-    // onNoteWave event
-    trk!(song).write_cc_on_note_wave(start_pos);
-    // write event
-    trk!(song).events.push(event);
-    trk!(song).timepos += notelen;
-}
-
-fn exec_rest(song: &mut Song, t: &Token) {
-    let trk = &mut song.tracks[song.cur_track];
-    let data_note_len = t.data[0].to_s();
-    let notelen = calc_length(&data_note_len, song.timebase, trk.length);
-    trk.timepos += notelen * t.value_i;
-}
-
-fn exec_voice(song: &mut Song, t: &Token) {
-    // voice no
-    let args = exec_args(song, t.children.as_ref().unwrap_or(&vec![]));
-    let no = if args.len() >= 1 { args[0].to_i() } else { 1 };
-    let no = value_range(1, no, 128) - 1;
-    let bank_msb = if args.len() >= 2 { args[1].to_i() } else { 0 };
-    let bank_lsb = if args.len() >= 3 { args[2].to_i() } else { 0 };
-    trk!(song).program_change = no + 1;
-    // bank ?
-    if args.len() == 1 {
-        song.add_event(Event::voice(trk!(song).timepos, trk!(song).channel, no));
-    } else {
-        song.add_event(Event::cc(trk!(song).timepos, trk!(song).channel, 0x00, bank_msb)); // msb
-        song.add_event(Event::cc(trk!(song).timepos, trk!(song).channel, 0x20, bank_lsb)); // lsb
-        song.add_event(Event::voice(trk!(song).timepos, trk!(song).channel, no));
-        // println!("voice: no={}, bank_msb={}, bank_lsb={}", no, bank_msb, bank_lsb);
-    }
-}
-
-fn exec_decres(song: &mut Song, t: &Token) {
-    let mut len_s = t.data[0].to_s();
-    if len_s == "" { len_s = "1".to_string(); }
-    let v1 = t.data[1].to_i();
-    let v2 = t.data[2].to_i();
-    let len = calc_length(&len_s, song.timebase, trk!(song).length);
-    let ia = vec![v1, v2, len];
-    // write EP
-    trk!(song).write_cc_on_time(11, ia);
 }
 
 pub fn value_range(min_v: isize, value: isize, max_v: isize) -> isize {
